feat: change state in switch story switch story의 상태를 바꿉니다
